Add support for Vault cluster and re-factor.
This commit adds support for a Vault cluster with an ability to initialize and
automatically unseal cluster nodes. Additionally, contains a lot of work on
re-factorings and clean-ups.

+++ packer/scripts/common/nodejs.sh
@@ -21,8 +21,10 @@ if [[ -n $NODEJS_VERSION ]]; then
     NODEJS_PACKAGE=$(printf 'nodejs=%s*' "$NODEJS_VERSION")
 fi
 
-wget -O "${NODEJS_FILES}/nodesource.key" \
-    https://deb.nodesource.com/gpgkey/nodesource.gpg.key
+if [[ ! -f  "${NODEJS_FILES}/nodesource.key" ]]; then
+    wget -O "${NODEJS_FILES}/nodesource.key" \
+        https://deb.nodesource.com/gpgkey/nodesource.gpg.key
+fi
 
 apt-key add "${NODEJS_FILES}/nodesource.key"
